+++ src/chains/Unichain/BridgeSettler.sol
@@ -32,12 +32,9 @@ contract UnichainBridgeSettler is BridgeSettler, Across, Mayan, StargateV2 {
         } else if (action == uint32(IBridgeSettlerActions.BRIDGE_NATIVE_TO_MAYAN.selector)) {
             (address forwarder, bytes memory protocolAndData) = abi.decode(data, (address, bytes));
             bridgeNativeToMayan(forwarder, protocolAndData);
-        } else if (action == uint32(IBridgeSettlerActions.BRIDGE_ERC20_TO_STARGATE_V2.selector)) {
+        } else if (action == uint32(IBridgeSettlerActions.BRIDGE_TO_STARGATE_V2.selector)) {
             (IERC20 token, address pool, bytes memory sendData) = abi.decode(data, (IERC20, address, bytes));
-            bridgeERC20ToStargateV2(token, pool, sendData);
-        } else if (action == uint32(IBridgeSettlerActions.BRIDGE_NATIVE_TO_STARGATE_V2.selector)) {
-            (address pool, uint256 destinationGas, bytes memory sendData) = abi.decode(data, (address, uint256, bytes));
-            bridgeNativeToStargateV2(pool, destinationGas, sendData);
+            bridgeToStargateV2(token, pool, sendData);
         } else {
             return false;
         }