+++ test/integration/StargateV2.t.sol
@@ -7,53 +7,13 @@ import {BridgeSettlerIntegrationTest} from "./BridgeSettler.t.sol";
 import {ALLOWANCE_HOLDER} from "src/allowanceholder/IAllowanceHolder.sol";
 import {IBridgeSettlerActions} from "src/bridge/IBridgeSettlerActions.sol";
 import {ArbitrumBridgeSettler} from "src/chains/Arbitrum/BridgeSettler.sol";
-
-interface IStargateV2 {
-    event OFTSent(
-        bytes32 indexed guid, uint32 dstEid, address indexed fromAddress, uint256 amountSentLD, uint256 amountReceivedLD
-    );
-
-    struct SendParam {
-        uint32 dstEid;
-        bytes32 to;
-        uint256 amountLD;
-        uint256 minAmountLD;
-        bytes extraOptions;
-        bytes composeMsg;
-        bytes oftCmd;
-    }
-
-    struct MessagingFee {
-        uint256 nativeFee;
-        uint256 lzTokenFee;
-    }
-
-    struct OFTLimit {
-        uint256 minAmountLD;
-        uint256 maxAmountLD;
-    }
-
-    struct OFTReceipt {
-        uint256 amountSentLD;
-        uint256 amountReceivedLD;
-    }
-
-    struct OFTFeeDetail {
-        int256 feeAmountLD;
-        string description;
-    }
-
-    function sendToken(SendParam memory SendParam, MessagingFee memory messagingFee, address refundAddress) external;
-
-    function quoteOFT(SendParam calldata _sendParam)
-        external
-        view
-        returns (OFTLimit memory, OFTFeeDetail[] memory oftFeeDetails, OFTReceipt memory);
-
-    function quoteSend(SendParam calldata _sendParam, bool _payInLzToken) external view returns (MessagingFee memory);
-}
+import {IStargateV2, IOFT, ETH} from "src/core/StargateV2.sol";
+import {ActionDataBuilder} from "../utils/ActionDataBuilder.sol";
+import {LibBytes} from "../utils/LibBytes.sol";
 
 contract StargateV2Test is BridgeSettlerIntegrationTest {
+    using LibBytes for bytes;
+
     address pool;
 
     receive() external payable {}
@@ -64,9 +24,9 @@ contract StargateV2Test is BridgeSettlerIntegrationTest {
 
     function _prepareSendToken(uint256 amount)
         internal
-        returns (IStargateV2.SendParam memory sendParam, IStargateV2.MessagingFee memory messagingFee, uint256 fee)
+        returns (IOFT.SendParam memory sendParam, IOFT.MessagingFee memory messagingFee, uint256 fee)
     {
-        sendParam = IStargateV2.SendParam({
+        sendParam = IOFT.SendParam({
             dstEid: uint32(30110), // ARBITRUM
             to: bytes32(uint256(uint160(makeAddr("recipient")))),
             amountLD: amount,
@@ -76,10 +36,10 @@ contract StargateV2Test is BridgeSettlerIntegrationTest {
             oftCmd: bytes("") // TAXI mode
         });
 
-        (,, IStargateV2.OFTReceipt memory receipt) = IStargateV2(pool).quoteOFT(sendParam);
+        (,, IOFT.OFTReceipt memory receipt) = IOFT(pool).quoteOFT(sendParam);
         sendParam.minAmountLD = receipt.amountReceivedLD;
 
-        messagingFee = IStargateV2(pool).quoteSend(sendParam, false);
+        messagingFee = IOFT(pool).quoteSend(sendParam, false);
         fee = messagingFee.nativeFee;
     }
 
@@ -87,31 +47,33 @@ contract StargateV2Test is BridgeSettlerIntegrationTest {
         // native pool
         pool = 0x77b2043768d28E9C9aB44E1aBfC95944bcE57931;
         uint256 amount = 1 ether;
-        uint256 extraGas = 10;
 
-        (IStargateV2.SendParam memory sendParam, IStargateV2.MessagingFee memory messagingFee, uint256 fee) =
+        (IOFT.SendParam memory sendParam, IOFT.MessagingFee memory messagingFee, uint256 fee) =
             _prepareSendToken(amount);
 
         sendParam.amountLD = 0; // send 0 to let settler inject the value
-        bytes[] memory bridgeActions = new bytes[](1);
-        bridgeActions[0] = abi.encodeCall(
-            IBridgeSettlerActions.BRIDGE_NATIVE_TO_STARGATE_V2,
-            (pool, extraGas, abi.encode(sendParam, messagingFee, address(this)))
+        bytes[] memory bridgeActions = ActionDataBuilder.build(
+            abi.encodeCall(
+                IBridgeSettlerActions.BRIDGE_TO_STARGATE_V2,
+                (address(ETH), pool, abi.encode(sendParam, messagingFee, address(this)))
+            )
         );
-        sendParam.amountLD = amount;
+        uint256 excess = 10;
+        sendParam.amountLD = amount + excess;
 
-        deal(address(this), amount + extraGas + fee);
+        deal(address(bridgeSettler), excess);
+        deal(address(this), amount + fee);
         uint256 balanceBefore = address(pool).balance;
         vm.expectCall(
             pool,
-            amount + extraGas + fee,
+            amount + excess + fee,
             abi.encodeCall(IStargateV2.sendToken, (sendParam, messagingFee, address(this)))
         );
-        bridgeSettler.execute{value: amount + extraGas + fee}(bridgeActions, bytes32(0));
+        bridgeSettler.execute{value: amount + fee}(bridgeActions, bytes32(0));
         uint256 balanceAfter = address(pool).balance;
 
         assertEq(balanceAfter - balanceBefore, amount, "Assets were not received");
-        assertEq(address(this).balance, extraGas, "ExtraGas was not returned");
+        assertEq(address(this).balance, excess, "Excess was not returned");
     }
 
     function testBridgeERC20() public {
@@ -123,26 +85,16 @@ contract StargateV2Test is BridgeSettlerIntegrationTest {
         deal(address(token), address(this), amount);
         token.approve(address(ALLOWANCE_HOLDER), amount);
 
-        (IStargateV2.SendParam memory sendParam, IStargateV2.MessagingFee memory messagingFee, uint256 fee) =
+        (IOFT.SendParam memory sendParam, IOFT.MessagingFee memory messagingFee, uint256 fee) =
             _prepareSendToken(amount);
 
-        bytes[] memory bridgeActions = new bytes[](2);
-        bridgeActions[0] = abi.encodeCall(
-            IBridgeSettlerActions.TRANSFER_FROM,
-            (
-                address(bridgeSettler),
-                ISignatureTransfer.PermitTransferFrom({
-                    permitted: ISignatureTransfer.TokenPermissions({token: address(token), amount: amount}),
-                    nonce: 0,
-                    deadline: block.timestamp
-                }),
-                bytes("")
-            )
-        );
         sendParam.amountLD = 0; // send 0 to let settler inject the value
-        bridgeActions[1] = abi.encodeCall(
-            IBridgeSettlerActions.BRIDGE_ERC20_TO_STARGATE_V2,
-            (address(token), pool, abi.encode(sendParam, messagingFee, address(this)))
+        bytes[] memory bridgeActions = ActionDataBuilder.build(
+            _getDefaultTransferFrom(address(token), amount),
+            abi.encodeCall(
+                IBridgeSettlerActions.BRIDGE_TO_STARGATE_V2,
+                (address(token), pool, abi.encode(sendParam, messagingFee, address(this)))
+            )
         );
         sendParam.amountLD = amount;